+++ src/config.js
@@ -4,7 +4,7 @@ import {
   getStatusBarHeight,
   getBottomSpace,
 } from 'react-native-iphone-x-helper';
-import { CART_PAYMENT_STATUS } from './constants/cart/types';
+import {CART_PAYMENT_STATUS, DELIVERY_STATUS_CODE} from './constants/cart';
 
 class Config {
   constructor() {
@@ -12,7 +12,7 @@ class Config {
   }
 
   get tagVersion() {
-    return 'r11.5.1';
+    return 'r11.5.3';
   }
 
   get appName() {
@@ -33,8 +33,8 @@ class Config {
 
   get voucherModule() {
     return {
-      appKey: 'abahaglobalkeywoxBg',
-      secretKey: 'dyTUabahaglobalkeywoxBg4OA1mh',
+      appKey: 'dieuthuykeyYhloT',
+      secretKey: '3SOrdieuthuykeyYhloTV6hCQ2',
     };
   }
 
@@ -84,11 +84,24 @@ class Config {
         warning: '#ffd166',
         success: '#06d6a0',
         info: '#118ab2',
-        other: '#073b4c'
+        other: '#073b4c',
       },
       cartType: {
         cart: '#B0C0F0',
-        dropship: '#FF9F1C'
+        dropship: '#FF9F1C',
+      },
+      delivery: {
+        [DELIVERY_STATUS_CODE.CANCEL]: '#DED9E2',
+        [DELIVERY_STATUS_CODE.SYSTEM_RECEIVED_ORDER]: '#9FA4C4',
+        [DELIVERY_STATUS_CODE.SHIPPER_RECEIVED_ORDER]: '#95AFBA',
+        [DELIVERY_STATUS_CODE.SHIPPER_GETTING_PACKAGES]: '#0F7173',
+        [DELIVERY_STATUS_CODE.FAIL_TO_GETTING_PACKAGES]: '#EF798A',
+        [DELIVERY_STATUS_CODE.DELIVERING]: '#ffd166',
+        [DELIVERY_STATUS_CODE.DELIVERY_SUCCESS]: '#06d6a0',
+        [DELIVERY_STATUS_CODE.DELIVERY_FAIL]: '#ef476f',
+        [DELIVERY_STATUS_CODE.ORDER_RETURNING]: '#92817A',
+        [DELIVERY_STATUS_CODE.ORDER_RETURNED]: '#6C7D47',
+        [DELIVERY_STATUS_CODE.FOR_CONTROL]: '#17A398',
       },
       orderStatus: {
         [CART_STATUS_CANCEL_1]: '#ef476f',
@@ -260,7 +273,7 @@ class Config {
       commissionIncomeStatement: 'commissionIncomeStatement',
 
       // Gamification
-      lotteryGame: 'lotteryGame'
+      lotteryGame: 'lotteryGame',
     };
   }
 }